Fix timestep dtype in load_inputs: use bfloat16 not Long
HunyuanVideoConditionEmbedding.forward does timesteps_proj.to(dtype=timestep.dtype),
so passing timestep as torch.long incorrectly converts the float sinusoidal projection
to integer, causing a dtype mismatch in the linear embedder. Use bfloat16 instead.

+++ hunyuan_video_1_5_i2v_720p_gguf/pytorch/loader.py
@@ -330,7 +330,9 @@ class ModelLoader(ForgeModel):
             config.image_embed_dim,
             dtype=dtype,
         )
-        timestep = torch.tensor([500], dtype=torch.long).expand(batch_size)
+        # HunyuanVideoConditionEmbedding.forward does timesteps_proj.to(dtype=timestep.dtype),
+        # so timestep must be float (not Long) to avoid converting sinusoidal projection to int.
+        timestep = torch.tensor([500], dtype=dtype).expand(batch_size)
 
         return {
             "hidden_states": hidden_states,
